+++ src/ast/entity/integer-value.ts
@@ -1,6 +1,6 @@
-import { Value } from '@/ast/model/value';
+import { Value } from '@/ast/entity/value';
 
-export class FloatValue extends Value {
+export class IntegerValue extends Value {
   private value: number;
 
   constructor(value: string | number) {
